# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your API Title",
        "description": "Your API Description",
        "contact": {
            "name": "Your Name",
            "email": "you@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/users": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "Get all users",
                "description": "Endpoint to retrieve a list of all users.",
                "operationId": "953fbe16ca028c3d65368ea6ce5a1ba4",
                "responses": {
                    "200": {
                        "description": "List of users"
                    }
                }
            },
            "post": {
                "tags": [
                    "users"
                ],
                "summary": "Create a new users",
                "description": "Endpoint to create a new users.",
                "operationId": "592819a0265360b2014512d6dbfaf0e7",
                "requestBody": {
                    "description": "Request Body Description",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "The name of the user",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "The email of the user",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "The password pf the user account",
                                        "type": "string"
                                    },
                                    "role": {
                                        "description": "The role of the user",
                                        "type": "string"
                                    },
                                    "id": {
                                        "description": "The roles to which the notification is targeted",
                                        "type": "integer"
                                    },
                                    "created_at": {
                                        "description": "The roles to which the notification is targeted",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "updated_at": {
                                        "description": "The roles to which the notification is targeted",
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Notification created successfully"
                    },
                    "400": {
                        "description": "Bad request, validation failed"
                    }
                }
            }
        },
        "/api/users/{user_id}": {
            "put": {
                "tags": [
                    "users"
                ],
                "summary": "Update user details",
                "description": "Endpoint to update user details.",
                "operationId": "c64f78766f4915182147dd62bcc1e3f3",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "ID of the user to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Request Body Description",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "The name of the user",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "The email of the user",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "The password of the user account",
                                        "type": "string"
                                    },
                                    "role": {
                                        "description": "The role of the user (e.g., administrator, manager, user)",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User details updated successfully"
                    },
                    "400": {
                        "description": "Bad request, validation failed"
                    }
                }
            },
            "delete": {
                "tags": [
                    "users"
                ],
                "summary": "Delete a user",
                "description": "Endpoint to delete a user by ID.",
                "operationId": "8813fccf55a87ec7b1237ecdee00419f",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "ID of the user to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "user deleted successfully"
                    },
                    "404": {
                        "description": "user not found"
                    }
                }
            }
        }
    }
}